feito card reutilizavel
