+++ labe-food/src/components/Feed/Feed.js
@@ -1,16 +1,26 @@
 import { CardContent, CardMedia, Typography } from "@mui/material";
-import { useContext } from "react";
+import { useContext, useState } from "react";
 import { GlobalStateContext } from "../../global/GlobalStateContext";
 import { ButtonNoStyle } from "../../pages/Home/styled-home";
 import { CardStyled, DivStyled, BodyContainer } from "./styled-feed";
 
-const Feed = () => {
+const Feed = ({isSearch}) => {
     const {restaurants, isLoading, error} = useContext(GlobalStateContext);
+    const [inputValue, setValue] = useState("")
 
-    console.log(restaurants);
+    const onChangeValue = (event) => {
+        setValue(event.target.value)
+    }
 
-    const renderRestaurants = restaurants && restaurants.map((restaurant) => {
-        console.log(restaurant.name)
+    const renderRestaurants = restaurants && restaurants.filter(restaurantes => {
+        if(isSearch)
+        if(!inputValue){
+            return false
+        }
+          return  restaurantes.name.toLowerCase().includes(inputValue.toLowerCase())
+        
+        })
+    .map((restaurant) => {
         return (
             <div>
                 <ButtonNoStyle>
@@ -18,7 +28,6 @@ const Feed = () => {
 
                         <CardMedia
                             component="img"
-                            // sx={{ maxWidth: 450}}
                             height="150"
                             image={restaurant.logoUrl}
                             alt="Foto do Restaurante"
@@ -39,6 +48,11 @@ const Feed = () => {
     })
     return (
         <DivStyled>
+
+                {isSearch ? <input
+                onChange={onChangeValue}
+                value={inputValue}
+            /> : <></>}
             {renderRestaurants}
             {!isLoading && error && <p>Deu um erro. Tente novamente.</p>}
         </DivStyled>
